# swagger/swagger.json
{"swagger":"2.0","info":{"title":"Carousell Challenge","description":"Carousell Challenge.","version":"0.1"},"schemes":["http","https"],"consumes":["application/json"],"produces":["application/json"],"paths":{"/api/v1/topics":{"get":{"tags":["topic"],"summary":"show topic","description":"Get top topics","operationId":"topic#show","produces":["text/plain"],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}},"schemes":["http","https"]},"post":{"tags":["topic"],"summary":"create topic","description":"Create new topic","operationId":"topic#create","produces":["text/plain"],"parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/CreateTopicPayload"}}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}},"schemes":["http","https"]}},"/api/v1/topics/{topicID}":{"delete":{"tags":["topic"],"summary":"delete topic","description":"remove the topic","operationId":"topic#delete","produces":["text/plain"],"parameters":[{"name":"topicID","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}},"schemes":["http","https"]},"patch":{"tags":["topic"],"summary":"vote topic","description":"upvote/downvote cluster","operationId":"topic#vote","produces":["text/plain"],"parameters":[{"name":"topicID","in":"path","required":true,"type":"string"},{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/VoteTopicPayload"}}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}},"schemes":["http","https"]}},"/api/v1/users":{"get":{"tags":["user"],"summary":"show user","description":"Get all users","operationId":"user#show","produces":["text/plain"],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}},"schemes":["http","https"]},"post":{"tags":["user"],"summary":"create user","description":"Create new user","operationId":"user#create","produces":["text/plain"],"parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/CreateUserPayload"}}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}},"schemes":["http","https"]}},"/api/v1/users/{userID}":{"delete":{"tags":["user"],"summary":"remove user","description":"remove user","operationId":"user#remove","produces":["text/plain"],"parameters":[{"name":"userID","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}},"schemes":["http","https"]}},"/api/v1/users/{userName}":{"get":{"tags":["user"],"summary":"get user","description":"Get user details by name","operationId":"user#get","produces":["text/plain"],"parameters":[{"name":"userName","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}},"schemes":["http","https"]}},"/swagger.json":{"get":{"summary":"Download swagger/swagger.json","operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http","https"]}},"/{filepath}":{"get":{"summary":"Download public/","operationId":"public#/*filepath","parameters":[{"name":"filepath","in":"path","description":"Relative file path","required":true,"type":"string"}],"responses":{"200":{"description":"File downloaded","schema":{"type":"file"}},"404":{"description":"File not found","schema":{"$ref":"#/definitions/error"}}},"schemes":["http","https"]}}},"definitions":{"CreateTopicPayload":{"title":"CreateTopicPayload","type":"object","properties":{"content":{"type":"string","description":"Content of the topic","example":"Non non tempora id consequatur dolor."},"userID":{"type":"string","description":"Creator of this topic","example":"Esse explicabo perspiciatis quo laudantium."}},"example":{"content":"Non non tempora id consequatur dolor.","userID":"Esse explicabo perspiciatis quo laudantium."},"required":["content","userID"]},"CreateUserPayload":{"title":"CreateUserPayload","type":"object","properties":{"country":{"type":"string","description":"country of the user","example":"Qui iste et."},"name":{"type":"string","description":"name of the user","example":"Ut qui et quia recusandae voluptatem."}},"example":{"country":"Qui iste et.","name":"Ut qui et quia recusandae voluptatem."},"required":["name","country"]},"VoteTopicPayload":{"title":"VoteTopicPayload","type":"object","properties":{"userID":{"type":"string","description":"user id","example":"Nihil voluptates dolorem."},"vote":{"type":"string","description":"upvote/downvote topic","example":"down","enum":["up","down"]}},"example":{"userID":"Nihil voluptates dolorem.","vote":"down"},"required":["userID","vote"]},"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}}}